Adicionado castin/float para conversão explícita

+++ media_nota.c
@@ -16,7 +16,7 @@ int main()
     printf("Digite sua  terceira nota:\n");
     scanf("%f", &nota3);
 
-    media = (nota1+nota2+nota3)/3;
+    media = (float)(nota1+nota2+nota3)/3;
 
     printf("A média é %.2f\n", media);
